Add GitHub action that updates policy data sources
Introduce a new `update-policy` workflow that is run only manually.
The action receives as input the name of the GitHub repository that
holds the policy, plus the tag associate with the release event of
that policy.

The action downloads the `hub.json` file from the repository mentioned
by the input variables.

The action then replaces the special `__TAG__` string inside of the
`hub.json` file with the value of the `tag` it received at invocation
time.

Finally, the action creates a PR against this repository with the
changes done to the file.